requirements.txt
crewai>=0.12.2
python-dotenv>=1.0.0
langchain>=0.1.0
google-generativeai>=0.3.2
openai>=1.12.0
requests>=2.31.0
python-dateutil>=2.8.2
typing-extensions>=4.9.0 